getOneDay

+++ src/main/java/com/example/demo/entity/TestElectric.java
@@ -14,12 +14,12 @@ import lombok.Data;
 public class TestElectric {
 
     //最大用电量
-    private Long MaxElectric;
+    private Double MaxElectric;
     //最小用电量
-    private Long MinElectric;
+    private Double MinElectric;
     //总共用电量
-    private Long SumElectric;
+    private Double SumElectric;
     //平均用电量
-    private Long AvgElectric;
+    private Double AvgElectric;
 
 }
